Changed position of permissions from Page_Qr to Login
